backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.0
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
openai>=1.3.7
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
openpyxl==3.1.2
pandas==2.1.4
mistralai==0.0.12
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3 